requirements.txt
fastapi
uvicorn
python-dotenv
pandas
selenium
pydantic
supabase